pyproject.toml
[project]
name = "pr-group7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.4.9",
    "keras>=3.11.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "tensorflow",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]
